+++ crates/arroyo-state-protocol/src/gc.rs
@@ -27,7 +27,7 @@ where
     let history =
         collect_history_and_clean_checkpoint_files(store, paths, head, new_min_epoch).await?;
 
-    for c in history {
+    for c in history.iter().rev() {
         debug!(
             generation = c.generation.0,
             epoch = c.epoch.0,
@@ -88,9 +88,10 @@ where
             }
             .into());
         }
-        history.push(owner);
 
-        if manifest.epoch < new_min_epoch.0 {
+
+        if manifest.epoch < *new_min_epoch {
+            history.push(owner);
             if let Err(e) = clean_checkpoint(store, paths, &checkpoint_ref, &manifest).await {
                 warn!(
                     checkpoint = checkpoint_ref.as_str(),
@@ -193,263 +194,4 @@ fn table_checkpoint_data_files(
     }
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::store::tests::MemoryProtocolStore;
-    use crate::store::{put_json, put_protobuf, read_protobuf};
-    use crate::types::{CommittedMarker, EpochRecord, PROTOCOL_VERSION};
-    use arroyo_rpc::grpc::rpc::{
-        ExpiringKeyedTimeTableCheckpointMetadata, OperatorCheckpointMetadata, OperatorMetadata,
-    };
-    use arroyo_types::{JobId, PipelineId};
-
-    fn paths() -> ProtocolPaths {
-        ProtocolPaths::new(PipelineId::new("P"), JobId::new("J"))
-    }
-
-    fn checkpoint_ref(paths: &ProtocolPaths, epoch: u64) -> CheckpointRef {
-        paths.checkpoint_manifest(Generation(1), Epoch(epoch))
-    }
-
-    fn data_ref(epoch: u64) -> CheckpointRef {
-        CheckpointRef::new(format!(
-            "P/J/generations/1/checkpoints/checkpoint-{epoch:07}/operator-op/table-table-000"
-        ))
-        .unwrap()
-    }
-
-    fn global_operator(files: Vec<CheckpointRef>) -> OperatorCheckpointMetadata {
-        OperatorCheckpointMetadata {
-            operator_metadata: Some(OperatorMetadata {
-                job_id: "J".to_string(),
-                operator_id: "op".to_string(),
-                epoch: 0,
-                min_watermark: None,
-                max_watermark: None,
-                parallelism: 1,
-            }),
-            start_time: 0,
-            finish_time: 0,
-            table_checkpoint_metadata: [(
-                "table".to_string(),
-                TableCheckpointMetadata {
-                    table_type: TableEnum::GlobalKeyValue.into(),
-                    data: GlobalKeyedTableTaskCheckpointMetadata {
-                        files: files.into_iter().map(|file| file.to_string()).collect(),
-                        commit_data_by_subtask: Default::default(),
-                    }
-                    .encode_to_vec(),
-                },
-            )]
-            .into(),
-            table_configs: Default::default(),
-        }
-    }
-
-    fn checkpoint(
-        paths: &ProtocolPaths,
-        epoch: u64,
-        parent: Option<u64>,
-        operators: Vec<OperatorCheckpointMetadata>,
-    ) -> CheckpointManifest {
-        CheckpointManifest {
-            pipeline_id: "P".to_string(),
-            job_id: "J".to_string(),
-            generation: 1,
-            epoch,
-            min_epoch: epoch,
-            start_time: 0,
-            finish_time: 0,
-            needs_commit: false,
-            operators,
-            parent_checkpoint_ref: parent.map(|epoch| checkpoint_ref(paths, epoch).to_string()),
-        }
-    }
-
-    async fn write_checkpoint(
-        store: &MemoryProtocolStore,
-        paths: &ProtocolPaths,
-        epoch: u64,
-        parent: Option<u64>,
-        operators: Vec<OperatorCheckpointMetadata>,
-    ) {
-        let this_checkpoint_ref = checkpoint_ref(paths, epoch);
-        let checkpoint = checkpoint(paths, epoch, parent, operators);
-        put_protobuf(store, &this_checkpoint_ref, &checkpoint)
-            .await
-            .unwrap();
-
-        let parent_checkpoint_ref = parent.map(|epoch| checkpoint_ref(paths, epoch));
-        let epoch_record = EpochRecord {
-            version: PROTOCOL_VERSION,
-            pipeline_id: PipelineId::new("P"),
-            job_id: JobId::new("J"),
-            epoch: Epoch(epoch),
-            generation: Generation(1),
-            parent_checkpoint_ref,
-            checkpoint_ref: this_checkpoint_ref.clone(),
-            created_at_micros: 0,
-        };
-        put_json(store, &paths.epoch_record(Epoch(epoch)), &epoch_record)
-            .await
-            .unwrap();
-
-        let committed_marker = CommittedMarker::new(
-            PipelineId::new("P"),
-            JobId::new("J"),
-            Epoch(epoch),
-            Generation(1),
-            Generation(1),
-            this_checkpoint_ref,
-        );
-        put_json(
-            store,
-            &paths.committed_marker(Generation(1), Epoch(epoch)),
-            &committed_marker,
-        )
-        .await
-        .unwrap();
-    }
-
-    async fn exists(store: &MemoryProtocolStore, path: &CheckpointRef) -> bool {
-        store.read_bytes(path).await.unwrap().is_some()
-    }
-
-    #[tokio::test]
-    async fn cleanup_deletes_only_checkpoints_below_new_min_epoch() {
-        let store = MemoryProtocolStore::default();
-        let paths = paths();
-        let file1 = data_ref(1);
-        let file2 = data_ref(2);
-        let file3 = data_ref(3);
-
-        store.put_bytes(&file1, b"1".to_vec()).await.unwrap();
-        store.put_bytes(&file2, b"2".to_vec()).await.unwrap();
-        store.put_bytes(&file3, b"3".to_vec()).await.unwrap();
-        write_checkpoint(
-            &store,
-            &paths,
-            1,
-            None,
-            vec![global_operator(vec![file1.clone()])],
-        )
-        .await;
-        write_checkpoint(
-            &store,
-            &paths,
-            2,
-            Some(1),
-            vec![global_operator(vec![file2.clone()])],
-        )
-        .await;
-        write_checkpoint(
-            &store,
-            &paths,
-            3,
-            Some(2),
-            vec![global_operator(vec![file3.clone()])],
-        )
-        .await;
-
-        cleanup_leader_checkpoints(&store, &paths, checkpoint_ref(&paths, 3), Epoch(2))
-            .await
-            .unwrap();
-
-        assert!(!exists(&store, &file1).await);
-        assert!(!exists(&store, &paths.epoch_record(Epoch(1))).await);
-        assert!(!exists(&store, &paths.committed_marker(Generation(1), Epoch(1))).await);
-        assert!(
-            read_protobuf::<_, CheckpointManifest>(&store, &checkpoint_ref(&paths, 1))
-                .await
-                .unwrap()
-                .is_none()
-        );
-
-        assert!(exists(&store, &file2).await);
-        assert!(exists(&store, &file3).await);
-        assert!(
-            read_protobuf::<_, CheckpointManifest>(&store, &checkpoint_ref(&paths, 2))
-                .await
-                .unwrap()
-                .is_some()
-        );
-        assert!(
-            read_protobuf::<_, CheckpointManifest>(&store, &checkpoint_ref(&paths, 3))
-                .await
-                .unwrap()
-                .is_some()
-        );
-    }
-
-    #[tokio::test]
-    async fn cleanup_deletes_checkpoint_manifests_oldest_first() {
-        let store = MemoryProtocolStore::default();
-        let paths = paths();
-        write_checkpoint(&store, &paths, 1, None, vec![global_operator(vec![])]).await;
-        write_checkpoint(&store, &paths, 2, Some(1), vec![global_operator(vec![])]).await;
-        write_checkpoint(&store, &paths, 3, Some(2), vec![global_operator(vec![])]).await;
-
-        cleanup_leader_checkpoints(&store, &paths, checkpoint_ref(&paths, 3), Epoch(3))
-            .await
-            .unwrap();
-
-        let deleted = store.deleted_objects();
-        let manifest1 = checkpoint_ref(&paths, 1).to_string();
-        let manifest2 = checkpoint_ref(&paths, 2).to_string();
-        let manifest3 = checkpoint_ref(&paths, 3).to_string();
-        let manifest1_pos = deleted
-            .iter()
-            .position(|path| path == &manifest1)
-            .expect("epoch 1 manifest should be deleted");
-        let manifest2_pos = deleted
-            .iter()
-            .position(|path| path == &manifest2)
-            .expect("epoch 2 manifest should be deleted");
-
-        assert!(manifest1_pos < manifest2_pos);
-        assert!(!deleted.contains(&manifest3));
-    }
-
-    #[tokio::test]
-    async fn cleanup_stops_at_missing_parent_manifest() {
-        let store = MemoryProtocolStore::default();
-        let paths = paths();
-        write_checkpoint(&store, &paths, 3, Some(2), vec![global_operator(vec![])]).await;
-
-        cleanup_leader_checkpoints(&store, &paths, checkpoint_ref(&paths, 3), Epoch(2))
-            .await
-            .unwrap();
-
-        assert!(
-            read_protobuf::<_, CheckpointManifest>(&store, &checkpoint_ref(&paths, 3))
-                .await
-                .unwrap()
-                .is_some()
-        );
-        assert!(store.deleted_objects().is_empty());
-    }
-
-    #[tokio::test]
-    async fn cleanup_detects_checkpoint_parent_cycles() {
-        let store = MemoryProtocolStore::default();
-        let paths = paths();
-        write_checkpoint(&store, &paths, 1, Some(2), vec![global_operator(vec![])]).await;
-        write_checkpoint(&store, &paths, 2, Some(1), vec![global_operator(vec![])]).await;
-
-        let err = cleanup_leader_checkpoints(&store, &paths, checkpoint_ref(&paths, 2), Epoch(0))
-            .await
-            .unwrap_err();
-
-        assert!(matches!(
-            err,
-            StoreError::Protocol(ProtocolError::CheckpointCycle {
-                generation: Generation(1),
-                epoch: Epoch(2)
-            })
-        ));
-        assert!(store.deleted_objects().is_empty());
-    }
 }
