+++ crates/arroyo-state-protocol/src/lib.rs
@@ -93,6 +93,7 @@ impl ProtocolPaths {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::gc::cleanup_leader_checkpoints;
     use crate::resolve::{
         EpochClaimOutcome, ParentCheckpointStatus, ResolveDecision, ResolveFailure,
         resolve_candidate,
@@ -100,8 +101,8 @@ mod tests {
     use crate::state::{CheckpointState, derive_checkpoint_state};
     use crate::store::tests::MemoryProtocolStore;
     use crate::store::{
-        CreateResult, StoreError, create_json_if_not_exist, put_json, put_protobuf, read_json,
-        read_protobuf,
+        CreateResult, ProtocolStore, StoreError, create_json_if_not_exist, put_json,
+        put_protobuf, read_json, read_protobuf,
     };
     use crate::types::{
         CommittedMarker, CurrentGeneration, EpochRecord, GenerationManifest, ProtocolError,
@@ -113,8 +114,12 @@ mod tests {
         initialize_generation, mark_committed, prepare_commit, publish_checkpoint,
         resolve_generation_manifest,
     };
-    use arroyo_rpc::grpc::rpc::CheckpointManifest;
+    use arroyo_rpc::grpc::rpc::{
+        CheckpointManifest, GlobalKeyedTableTaskCheckpointMetadata, OperatorCheckpointMetadata,
+        OperatorMetadata, TableCheckpointMetadata, TableEnum,
+    };
     use arroyo_types::{JobId, PipelineId, from_micros};
+    use prost::Message;
     use std::time::SystemTime;
 
     fn checkpoint_ref(path: &str) -> CheckpointRef {
@@ -244,6 +249,203 @@ mod tests {
         )
         .await
         .unwrap();
+    }
+
+    fn data_ref(paths: &ProtocolPaths, epoch: u64) -> CheckpointRef {
+        checkpoint_ref(&format!(
+            "{}/operator-op/table-table-000",
+            paths.checkpoint_dir(Generation(1), Epoch(epoch))
+        ))
+    }
+
+    fn global_operator(files: Vec<CheckpointRef>) -> OperatorCheckpointMetadata {
+        OperatorCheckpointMetadata {
+            operator_metadata: Some(OperatorMetadata {
+                job_id: "J".to_string(),
+                operator_id: "op".to_string(),
+                epoch: 0,
+                min_watermark: None,
+                max_watermark: None,
+                parallelism: 1,
+            }),
+            start_time: 0,
+            finish_time: 0,
+            table_checkpoint_metadata: [(
+                "table".to_string(),
+                TableCheckpointMetadata {
+                    table_type: TableEnum::GlobalKeyValue.into(),
+                    data: GlobalKeyedTableTaskCheckpointMetadata {
+                        files: files.into_iter().map(|file| file.to_string()).collect(),
+                        commit_data_by_subtask: Default::default(),
+                    }
+                    .encode_to_vec(),
+                },
+            )]
+            .into(),
+            table_configs: Default::default(),
+        }
+    }
+
+    async fn write_gc_checkpoint(
+        store: &MemoryProtocolStore,
+        paths: &ProtocolPaths,
+        epoch: u64,
+        parent_epoch: Option<u64>,
+        operators: Vec<OperatorCheckpointMetadata>,
+    ) {
+        let checkpoint_ref = paths.checkpoint_manifest(Generation(1), Epoch(epoch));
+        let parent_checkpoint_ref =
+            parent_epoch.map(|epoch| paths.checkpoint_manifest(Generation(1), Epoch(epoch)));
+        let mut checkpoint =
+            checkpoint_for_generation(Generation(1), epoch, parent_checkpoint_ref, false);
+        checkpoint.operators = operators;
+        write_canonical_checkpoint(store, paths, &checkpoint_ref, &checkpoint).await;
+        put_json(
+            store,
+            &paths.committed_marker(Generation(1), Epoch(epoch)),
+            &committed_marker(checkpoint_ref, epoch),
+        )
+        .await
+        .unwrap();
+    }
+
+    async fn exists(store: &MemoryProtocolStore, path: &CheckpointRef) -> bool {
+        store.read_bytes(path).await.unwrap().is_some()
+    }
+
+    #[tokio::test]
+    async fn cleanup_deletes_only_checkpoints_below_new_min_epoch() {
+        let store = MemoryProtocolStore::default();
+        let paths = ProtocolPaths::new(PipelineId::new("P"), JobId::new("J"));
+        let file1 = data_ref(&paths, 1);
+        let file2 = data_ref(&paths, 2);
+        let file3 = data_ref(&paths, 3);
+        let checkpoint1_ref = paths.checkpoint_manifest(Generation(1), Epoch(1));
+        let checkpoint2_ref = paths.checkpoint_manifest(Generation(1), Epoch(2));
+        let checkpoint3_ref = paths.checkpoint_manifest(Generation(1), Epoch(3));
+
+        store.put_bytes(&file1, b"1".to_vec()).await.unwrap();
+        store.put_bytes(&file2, b"2".to_vec()).await.unwrap();
+        store.put_bytes(&file3, b"3".to_vec()).await.unwrap();
+        write_gc_checkpoint(
+            &store,
+            &paths,
+            1,
+            None,
+            vec![global_operator(vec![file1.clone()])],
+        )
+        .await;
+        write_gc_checkpoint(
+            &store,
+            &paths,
+            2,
+            Some(1),
+            vec![global_operator(vec![file2.clone()])],
+        )
+        .await;
+        write_gc_checkpoint(
+            &store,
+            &paths,
+            3,
+            Some(2),
+            vec![global_operator(vec![file3.clone()])],
+        )
+        .await;
+
+        cleanup_leader_checkpoints(&store, &paths, checkpoint3_ref.clone(), Epoch(2))
+            .await
+            .unwrap();
+
+        assert!(!exists(&store, &file1).await);
+        assert!(!exists(&store, &paths.epoch_record(Epoch(1))).await);
+        assert!(!exists(&store, &paths.committed_marker(Generation(1), Epoch(1))).await);
+        assert!(
+            read_protobuf::<_, CheckpointManifest>(&store, &checkpoint1_ref)
+                .await
+                .unwrap()
+                .is_none()
+        );
+
+        assert!(exists(&store, &file2).await);
+        assert!(exists(&store, &file3).await);
+        assert!(
+            read_protobuf::<_, CheckpointManifest>(&store, &checkpoint2_ref)
+                .await
+                .unwrap()
+                .is_some()
+        );
+        assert!(
+            read_protobuf::<_, CheckpointManifest>(&store, &checkpoint3_ref)
+                .await
+                .unwrap()
+                .is_some()
+        );
+    }
+
+    #[tokio::test]
+    async fn cleanup_deletes_checkpoint_manifests_oldest_first() {
+        let store = MemoryProtocolStore::default();
+        let paths = ProtocolPaths::new(PipelineId::new("P"), JobId::new("J"));
+        write_gc_checkpoint(&store, &paths, 1, None, vec![global_operator(vec![])]).await;
+        write_gc_checkpoint(&store, &paths, 2, Some(1), vec![global_operator(vec![])]).await;
+        write_gc_checkpoint(&store, &paths, 3, Some(2), vec![global_operator(vec![])]).await;
+
+        cleanup_leader_checkpoints(
+            &store,
+            &paths,
+            paths.checkpoint_manifest(Generation(1), Epoch(3)),
+            Epoch(3),
+        )
+        .await
+        .unwrap();
+
+        let deleted = store.deleted_objects();
+        let manifest1 = paths
+            .checkpoint_manifest(Generation(1), Epoch(1))
+            .to_string();
+        let manifest2 = paths
+            .checkpoint_manifest(Generation(1), Epoch(2))
+            .to_string();
+        let manifest3 = paths
+            .checkpoint_manifest(Generation(1), Epoch(3))
+            .to_string();
+        let manifest1_pos = deleted
+            .iter()
+            .position(|path| path == &manifest1)
+            .expect("epoch 1 manifest should be deleted");
+        let manifest2_pos = deleted
+            .iter()
+            .position(|path| path == &manifest2)
+            .expect("epoch 2 manifest should be deleted");
+
+        assert!(manifest1_pos < manifest2_pos);
+        assert!(!deleted.contains(&manifest3));
+    }
+
+    #[tokio::test]
+    async fn cleanup_detects_checkpoint_parent_cycles() {
+        let store = MemoryProtocolStore::default();
+        let paths = ProtocolPaths::new(PipelineId::new("P"), JobId::new("J"));
+        write_gc_checkpoint(&store, &paths, 1, Some(2), vec![global_operator(vec![])]).await;
+        write_gc_checkpoint(&store, &paths, 2, Some(1), vec![global_operator(vec![])]).await;
+
+        let err = cleanup_leader_checkpoints(
+            &store,
+            &paths,
+            paths.checkpoint_manifest(Generation(1), Epoch(2)),
+            Epoch(0),
+        )
+        .await
+        .unwrap_err();
+
+        assert!(matches!(
+            err,
+            StoreError::Protocol(ProtocolError::CheckpointCycle {
+                generation: Generation(1),
+                epoch: Epoch(2)
+            })
+        ));
+        assert!(store.deleted_objects().is_empty());
     }
 
     #[tokio::test]
